Created the increment method in the StatusManager. The Application is pretty much working.

+++ src/main/java/edu/aws/pwgenerator/domain/Event.java
@@ -14,7 +14,7 @@ import javax.persistence.Id;
 public class Event {
 
     @Id
-    private Integer idevents;
+    private Long idevents;
     private String event;
     private String eventyear;
