+++ src/main/java/edu/aws/pwgenerator/domain/EventsRepository.java
@@ -2,6 +2,6 @@ package edu.aws.pwgenerator.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface EventsRepository extends CrudRepository<Event, Integer> {
+public interface EventsRepository extends CrudRepository<Event, Long> {
 
 }
