+++ src/main/java/edu/aws/pwgenerator/domain/NamesRepository.java
@@ -2,6 +2,6 @@ package edu.aws.pwgenerator.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface NamesRepository extends CrudRepository<Name, Integer> {
+public interface NamesRepository extends CrudRepository<Name, Long> {
 
 }
