+++ src/main/java/edu/aws/pwgenerator/domain/PlacesRepository.java
@@ -2,6 +2,6 @@ package edu.aws.pwgenerator.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlacesRepository extends CrudRepository<Place, Integer> {
+public interface PlacesRepository extends CrudRepository<Place, Long> {
 
 }
